Draw interactively a polygon and get it as Well Known Text

Draw interactively a polygon and get it as Well Known Text

Usage

eurobis_map_draw()

Value

a string with a polygon as Well Known Text

Details

Set options(verbose = TRUE) to display more information

Examples

if (FALSE) wkt <- eurobis_map_draw() # \dontrun{}