Interactive map with the regions on your viewer. Click to get the MRGID and more info

Interactive map with the regions on your viewer. Click to get the MRGID and more info

Usage

eurobis_map_regions(layer = "eez")

eurobis_map_regions_ecoregions()

eurobis_map_regions_eez()

eurobis_map_regions_iho()

eurobis_map_regions_eez_iho()

eurobis_map_regions_reportingareas()

Arguments

layer

The region or place type to be displayed. See details.

Value

A leaflet map with the regions that retrieves info by clicking

Details

To select a specific layer you have to provide the code for the layer. Or alternatively, use the helper of each layer: - eurobis_map_regions_ecoregions() = Marine Ecoregions of the World (MEOW) - eurobis_map_regions_eez() = International Hydrographic Office Sea Areas (IHO) - eurobis_map_regions_iho() = Exclusive Economic Zones (EEZ) - eurobis_map_regions_eez_iho() = Intersections of IHO and EEZ - eurobis_map_regions_reportingareas() = EMODnet Biology Reporting Areas

Marine Regions https://marineregions.org/ is the geographical backbone of EurOBIS. All the regions are included in the Marine Regions Gazetteer. You can see the full list of regions with eurobis_list_regions().

You can find more info about these layers in https://marineregions.org/sources.php

The EMODnet Biology Reporting Areas are not described there. The EMODnet Reporting Areas are modified from the MSDF European seas (https://www.eea.europa.eu/data-and-maps/data/europe-seas-1) to fit the purpose of reporting in the EMODnet project. More info in: https://marineregions.org/gazetteer.php?p=details&id=63392

Examples

if (FALSE) { # \dontrun{
eurobis_map_regions_ecoregions()
eurobis_map_regions_eez()
eurobis_map_regions_iho()
eurobis_map_regions_eez_iho()
eurobis_map_regions_reportingareas()
} # }